// src/app/api-client/model/changePlanDTO.ts
/**
 * TeachingBACKEND API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BillingInterval } from './billingInterval';


export interface ChangePlanDTO { 
    newPlanId: string;
    billingInterval: BillingInterval;
    prorate?: boolean;
    reason?: string | null;
}
export namespace ChangePlanDTO {
}